Python3

+++ sebastian/core/notes.py
@@ -22,7 +22,7 @@ def double_flat(val):
     return -10 > val > -18
 
 def modifiers(val):
-    return ((val + 3) - ((val + 3) % 7)) / 7
+    return int(((val + 3) - ((val + 3) % 7)) / 7)
 
 def mod_interval(mod):
     return 7 * mod
